+++ native/BANG_FX/src/BANG_ClonerPiPL.r
@@ -22,7 +22,7 @@ resource 'PiPL' (16000) {
 #endif
 		AE_PiPL_Version { 2, 0 },
 		AE_Effect_Spec_Version { PF_PLUG_IN_VERSION, PF_PLUG_IN_SUBVERS },
-		AE_Effect_Version { 589825 /* 1.2.0 build 1 */ },
+		AE_Effect_Version { 720897 /* 1.6.0 build 1 */ },
 		AE_Effect_Info_Flags { 0 },
 		/* PF_OutFlag_DEEP_COLOR_AWARE | PF_OutFlag_I_EXPAND_BUFFER | PF_OutFlag_SEND_UPDATE_PARAMS_UI | PF_OutFlag_CUSTOM_UI */
 		AE_Effect_Global_OutFlags { 0x06008200 },